# FrontEnd/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Passluk",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Passluk",
      "email": "BreadExchange@gmail.com"
    },
    "license": {
      "name": "Licence BradExchange",
      "url": "https://BreadExchange.com"
    },
    "version": "2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/frontEnd/searchBakery": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "searchBakery",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BakeryNavBarResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/isOpen": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "isOpen",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getWeek": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getWeek",
        "parameters": [
          {
            "name": "idBakery",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          }
        }
      }
    },
...
    "/frontEnd/getRandomData": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getRandomData",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RandomDataBakeryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getPlaneAddress": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getPlaneAdd",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanAddBkery"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getInfo": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getCategory": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getCategory",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getBAkeryById": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getBAkeryById",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BakeryNavBarResponse"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/getAllCategory": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "getAllCategory",
        "parameters": [
          {
            "name": "owner",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCategoryFrontEnd"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/checkOwner": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "checkOwner",
        "parameters": [
          {
            "name": "idBack",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/frontEnd/bakery/{id}": {
      "get": {
        "tags": [
          "front-end-controller"
        ],
        "operationId": "bakeryDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BakeryfrontEndResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/searchByName": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "searchByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBakeryRegisterRequest"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "telNumber": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "provincia": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          }
        }
      },
      "Bakery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "url_picture": {
            "type": "string"
          },
          "url_backImg": {
            "type": "string"
          },
          "contactInfo": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "week": {
            "$ref": "#/components/schemas/Week"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "registrationDate": {
            "type": "string",
            "format": "date-time"
          },
          "openStatus": {
            "type": "boolean"
          },
          "abilitato": {
            "type": "boolean"
          }
        }
      },
      "BakeryFav": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "pictures": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
      "BakeryRegisterRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email_azz": {
            "type": "string"
          },
          "phone_azz": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "enable": {
            "type": "boolean"
          },
          "instagram": {
            "type": "string"
          }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsCart"
            }
          },
          "total": {
            "type": "number"
          },
          "idBac": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "owner": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "ContactInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          }
        }
      },
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "img": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "ItemsCart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "details": {
            "$ref": "#/components/schemas/Item"
          },
          "subTotal": {
            "type": "number"
          }
        }
      },
      "JwtToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "token": {
            "type": "string"
          },
          "revoked": {
            "type": "boolean"
          },
          "expired": {
            "type": "boolean"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "bakery": {
            "$ref": "#/components/schemas/Bakery"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemsCart"
            }
          },
          "accepted": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "url_picture": {
            "type": "string"
          },
          "url_BackImg": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "enabled": {
            "type": "boolean"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "fav": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BakeryFav"
            }
          },
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JwtToken"
            }
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "BAKERY_OWNER",
              "ADMIN"
            ]
          },
          "userName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "enable": {
            "type": "boolean",
            "writeOnly": true
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
      "Week": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lun": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "mar": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "mer": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "gio": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "ven": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "sab": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "dom": {
            "$ref": "#/components/schemas/WeekDay"
          }
        }
      },
      "WeekDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "morningOpening": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "morningClosing": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "eveningOpening": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "eveningClosing": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "closed": {
            "type": "boolean"
          },
          "allDayOpen": {
            "type": "boolean"
          }
        }
      },
...
      "BakeryNavBarResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          }
        }
      },
...
      "RandomDataBakeryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "PlanAddBkery": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          }
        }
      },
      "ListCategoryFrontEnd": {
        "type": "object",
        "properties": {
          "name_azz": {
            "type": "string"
          },
          "id_azz": {
            "type": "integer",
            "format": "int32"
          },
          "img_azz": {
            "type": "string"
          },
          "categorys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        }
      },
      "BakeryfrontEndResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "owner": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "Logo": {
            "type": "string"
          },
          "Back_img": {
            "type": "string"
          },
          "info": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "categorys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "logo": {
            "type": "string"
          },
          "back_img": {
            "type": "string"
          }
        }
      },
      "PageResponseBakeryRegisterRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BakeryRegisterRequest"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}